+++ templates/esm/routes/home.js
@@ -1,0 +1,11 @@
+import {Router} from 'express';
+const router = Router();
+
+/**
+ * Home page.
+ */
+router.get('/', (req, res) => {
+  res.render('home');
+});
+
+export default router;