+++ templates/esm/routes/login.js
@@ -1,0 +1,11 @@
+import {Router} from 'express';
+const router = Router();
+
+/**
+ * Login page.
+ */
+router.get('/', (req, res) => {
+  res.render('login');
+});
+
+export default router;